Acomodo de metodos HTTP de user, y creacion de modelos necesarios para flightplan

+++ app/repositories/user_repository.py
@@ -1,3 +1,5 @@
+import logging
+from math import e
 from typing import List
 from app.models import User
 from app import db
@@ -10,7 +12,32 @@ class UserRepository(CreateAbstractRepository, ReadAbstractRepository, DeleteAbs
         db.session.add(user)
         db.session.commit()
         return user
-   
+    
+    @staticmethod
+    def update(user: User, id: int) -> User:
+        entity = UserRepository.find(id)
+        
+        if entity is None:
+            return None
+        
+        entity.firstname = user.firstname
+        entity.lastname = user.lastname
+        entity.dni = user.dni
+        entity.email = user.email
+        entity.phone = user.phone
+        entity.address = user.address
+        entity.state = user.state
+        entity.city = user.city
+        entity.zipcode = user.zipcode
+        
+        if user.password is not None:
+            entity.password = user.password
+        
+        db.session.add(entity)
+        db.session.commit()
+        return entity
+        
+    
     @staticmethod   
     def find_all() -> List[User]:
         return User.query.all()
@@ -21,9 +48,19 @@ class UserRepository(CreateAbstractRepository, ReadAbstractRepository, DeleteAbs
         
     @staticmethod
     def find(id: int) -> User:
-        return User.query.get(id)
+        result = None
+        if id is not None:
+            try:
+                result = User.query.get(id)
+            except Exception as e:
+                logging.error(f'error getting user by id: {id}, {e}') 
+        return result
     
     @staticmethod
-    def delete(user: User) -> None:
-        db.session.delete(user)
-        db.session.commit()
+    def delete(id: int) -> None:
+        user = UserRepository.find(id)
+        if user:
+            db.session.delete(user)
+            db.session.commit()
+        else:
+            logging.error(f'error deleting user by id: {id}')
